Add files via upload
Created a new file called newUserFrame that is a jApplet so it can be attached to a web browser in the future. For testing purposes right now the main file of the project creates a jFrame for the applet to run in. Registration and login code now both connect to the aws database. Included the jsql driver jar file.